new projectile for godsip, some other architectural changes

+++ src/projectiles/WeaponSpec.hpp
@@ -3,7 +3,7 @@
 struct WeaponSpec {
   int speed = 0;
   float fireRate = 99.0f;
-  int size = 0;
+  float size = 0.0f;
   float lifetime = 0;
   int damage = 0;
 };
